[Untracked] Fix feesPerGas null check & bump up default gas fee multiplier (#4)
# Summary
When testing ARB-SEPOLIA there were errors running a token-transfer
since there were errors getting the gas fee. After looking closer, one
of the values being returned ```maxPriorityFeePerGas``` was 0. On the
next line we did a null check, but based on on JavaScript handles falsly
values it treated 0 as false 🙃

So I updated the comparison to explicitly check for null or undefined. I
also adjusted the calculations we did with
```adjustedMaxPriorityFeePerGas``` so we'd also scale up the value we
got from alchemy. Without scaling this up the gas was too low to get
accepted, kept getting ```replacement underpriced``` responses.

Also bumped up default gas fee to 1.2 multiplier. If at 1 most of the
time there are problems sending out a transaction.

## Detail

## Testing

## Documentation

---

**Requested Reviewers:** @mention